requirements.txt
PyQt5
pyinstaller
reportlab
arabic-reshaper
python-bidi
Pillow
watchdog
plyer
